# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/patient": {
            "get": {
                "description": "Bad Request Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Bad Request Patient",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestPatientsReponse"
                        }
                    }
                }
            }
        },
        "/patient/create": {
            "post": {
                "description": "Create a new Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Create Patient",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePatientsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePatientsReponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patient/{id}": {
            "get": {
                "description": "Show a Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Show a Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient identification",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ShowPatientsReponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient identification",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdatePatientsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdatePatientsReponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Delete Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient identification",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeletePatientsReponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients": {
            "get": {
                "description": "List Patients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "List Patients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ListPatientsReponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BadRequestPatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.PatientsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.CreatePatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.PatientsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.CreatePatientsRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nameTutor": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "handler.DeletePatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.PatientsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ListPatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.PatientsResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ShowPatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.PatientsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.UpdatePatientsReponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schemas.PatientsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.UpdatePatientsRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nameTutor": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "schemas.PatientsResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nameTutor": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        }
    }
}